requirements.txt
torch
torchvision
pillow
numpy
accelerate
safetensors
timm
peft==0.9.0
diffusers==0.25.1
huggingface-hub==0.25.1
transformers==4.38.0